refactor/가격 데이터타입 전부 int로 변경,db컬럼도 전부 int로 통일

+++ src/main/java/ko/dh/goot/order/domain/Order.java
@@ -20,7 +20,7 @@ public class Order {
     private Long orderId;
     private String userId;
     private String orderName;
-    private long totalAmount;
+    private int totalAmount;
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus;
     private String receiverName;
